Github Action: use newer node base image (buster instead of stretch)
The default node:14 image is based on Debian Stretch which has an older git version, which leads to the checkout action changing its behaviour: it downloads a git archive file and unpacks this, so it's no longer a git archive. As a result, Antora won't process that local directory as content source.